Små endringer
Statisticsfragment og service

+++ src/com/example/norserun/StatisticsFragment.java
@@ -116,8 +116,8 @@ public class StatisticsFragment extends SherlockFragment implements OnClickListe
 			Log.d(SQL_TAG, "Strings: " + nameFromDb + latFromDb + longtFromDb + timeFromDb);
 			
 			liste = StatisticsHelper.StringDeserializer(latFromDb, longtFromDb, timeFromDb);
-			speedText.setText(String.valueOf(StatisticsHelper.GetAverageSpeed(liste))+" Km/timen");
-			distanceText.setText(String.valueOf(StatisticsHelper.GetDistance(liste))+" meter");
+			speedText.setText(String.valueOf((float)StatisticsHelper.GetAverageSpeed(liste))+" Km/timen");
+			distanceText.setText(String.valueOf((float)StatisticsHelper.GetDistance(liste))+" meter");
 			titleText.setText(nameFromDb);
 //			Log.d("POSISJON", String.valueOf(StatisticsHelper.GetAverageSpeed(liste)));
 //			Log.d("POSISJON", String.valueOf(StatisticsHelper.GetDistance(liste)));
